Extract the Operator Inbox builders into a partial
Increment 2i of the endpoint-monolith decomposition. Moves the Operator Inbox
builder family (16 private static methods, BuildOperatorInboxAsync through
BuildOperatorInboxScopedId) out of src/Meridian.Ui.Shared/Endpoints/WorkstationEndpoints.cs
into a new sibling partial WorkstationEndpoints.OperatorInbox.cs.

Behavior-preserving verbatim relocation: WorkstationEndpoints is one partial
class, so the inline inbox route lambda keeps calling BuildOperatorInboxAsync
unchanged, and the moved BuildOperatorInboxScopedId /
AddReconciliationBreakWorkItemsAsync still reach the core-resident
NormalizeOperatorInboxToken and GetBreakQueueItemsAsync across the partial.
Closed call graph verified: the 16 methods have zero references in other
partials; the only entry is the single inline route registration in core.

Deliberately left in core (shared): the route lambda, GetTradingOperatorReadinessAsync
(shared with the trading payload), and NormalizeOperatorInboxToken + its
MaxOperatorInboxTokenLength const (shared with collateral).

Core WorkstationEndpoints.cs drops 5,976 -> 5,636 lines; ratchet baseline
retightened.
